third_party/blink/web_tests/external/wpt/css/css-borders/corner-shape/corner-shape-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property corner-top-left-shape value 'round'
  assert_equals: expected "superellipse(1)" but got "round"
[FAIL] Property corner-top-left-shape value 'scoop'
  assert_equals: expected "superellipse(-1)" but got "scoop"
[FAIL] Property corner-top-left-shape value 'superellipse(-infinity)'
  assert_equals: expected "superellipse(-infinity)" but got "notch"
[FAIL] Property corner-top-left-shape value 'superellipse(infinity)'
  assert_equals: expected "superellipse(infinity)" but got "square"
[FAIL] Property corner-top-left-shape value 'superellipse(1)'
  assert_equals: expected "superellipse(1)" but got "round"
[FAIL] Property corner-top-left-shape value 'superellipse(0)'
  assert_equals: expected "superellipse(0)" but got "bevel"
[FAIL] Property corner-top-left-shape value 'superellipse(2)'
  assert_equals: expected "superellipse(2)" but got "squircle"
[FAIL] Property corner-top-left-shape value 'superellipse( -1)'
  assert_equals: expected "superellipse(-1)" but got "scoop"
[FAIL] Property corner-top-right-shape value 'round'
  assert_equals: expected "superellipse(1)" but got "round"
[FAIL] Property corner-bottom-right-shape value 'scoop'
  assert_equals: expected "superellipse(-1)" but got "scoop"
[FAIL] Property corner-shape value 'superellipse(5) round'
  assert_equals: expected "superellipse(5) superellipse(1)" but got "superellipse(5) round"
[FAIL] Property corner-shape value 'round'
  assert_equals: expected "superellipse(1)" but got "round"
[FAIL] Property corner-shape value 'bevel superellipse(0.1) round squircle'
  assert_equals: expected "superellipse(0) superellipse(0.1) superellipse(1) superellipse(2)" but got "bevel superellipse(0.1) round squircle"
[FAIL] Property corner-shape value 'round round round round'
  assert_equals: expected "superellipse(1)" but got "round"
[FAIL] Property corner-shape value 'round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-shape value 'round scoop round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-shape value 'bevel superellipse(1)'
  assert_equals: expected "superellipse(0) superellipse(1)" but got "bevel round"
[FAIL] Property corner-shape value 'superellipse(-1) superellipse(3) square'
  assert_equals: expected "superellipse(-1) superellipse(3) superellipse(infinity)" but got "scoop superellipse(3) square"
[FAIL] Property corner-shape value 'superellipse(-1) superellipse(3) superellipse(0)'
  assert_equals: expected "superellipse(-1) superellipse(3) superellipse(0)" but got "scoop superellipse(3) bevel"
[FAIL] Property corner-shape value 'bevel superellipse(1) squircle     round'
  assert_equals: expected "superellipse(0) superellipse(1) superellipse(2)" but got "bevel round squircle"
[FAIL] Property corner-shape value 'superellipse(-1) superellipse(3) superellipse(0) superellipse(infinity)'
  assert_equals: expected "superellipse(-1) superellipse(3) superellipse(0) superellipse(infinity)" but got "scoop superellipse(3) bevel square"
[FAIL] Property corner-left-shape value 'round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-top-shape value 'round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-right-shape value 'superellipse(-1) superellipse(2)'
  assert_equals: expected "superellipse(-1) superellipse(2)" but got "scoop squircle"
[FAIL] Property corner-bottom-shape value 'scoop scoop'
  assert_equals: expected "superellipse(-1)" but got "scoop"
[FAIL] Property corner-inline-start-shape value 'round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-block-start-shape value 'round scoop'
  assert_equals: expected "superellipse(1) superellipse(-1)" but got "round scoop"
[FAIL] Property corner-inline-start-shape value 'superellipse(-1) superellipse(2)'
  assert_equals: expected "superellipse(-1) superellipse(2)" but got "scoop squircle"
[FAIL] Property corner-block-end-shape value 'scoop scoop'
  assert_equals: expected "superellipse(-1)" but got "scoop"
Harness: the test ran to completion.
